requirements/local.txt
-r base.txt

# Testing
# ------------------------------------------------------------------------------
mypy  # https://github.com/python/mypy
pytest  # https://github.com/pytest-dev/pytest
pytest-cov
pytest-django
pytest-dotenv
pytest-mock
pytest-pythonpath

# Code quality
# ------------------------------------------------------------------------------
flake8==6.1.0  # https://github.com/PyCQA/flake8
flake8-isort==6.0.0  # https://github.com/gforcada/flake8-isort
coverage==6.4.2  # https://github.com/nedbat/coveragepy
black==23.7.0  # https://github.com/psf/black
djlint==1.32.1  # https://github.com/Riverside-Healthcare/djLint
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pre-commit==3.3.3  # https://github.com/pre-commit/pre-commit
